DRTII-213 Tidying
